refactor: rename error recovery hook for clarity
